move/rename dataType, update dependency paths
